requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
jinja2==3.1.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0

SQLAlchemy==2.0.30
aiosqlite==0.20.0
alembic==1.13.2

pydantic==2.7.4
pydantic-settings==2.3.3

python-slugify==8.0.4
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0

requests==2.32.3
itsdangerous==2.2.0
httpx[http2]==0.27.0